Code cleanup and a bit of refactoring
