+++ frontend/src/credits/components/AnalystRecommendationHeader.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const AnalystRecommendationHeader = () => (
+  <div className="p-2">
+    <h4>Analyst Recommendation</h4>
+    <ul className="mt-2 mb-0 pl-4">
+      <li>View sales not eligible for credits or overridden by the analyst from system default</li>
+    </ul>
+  </div>
+);
+
+export default AnalystRecommendationHeader;
