+++ frontend/src/credits/components/CreditRequestDetailsPage.js
@@ -403,7 +403,6 @@ CreditRequestDetailsPage.propTypes = {
   submission: PropTypes.shape().isRequired,
   uploadDate: PropTypes.string.isRequired,
   user: CustomPropTypes.user.isRequired,
-  files: PropTypes.arrayOf(PropTypes.shape()),
 };
 
 export default CreditRequestDetailsPage;
